requirements.txt
# Core dependencies
python-dotenv>=1.0.0
click>=8.1.0
requests>=2.31.0
sqlalchemy>=2.0.0
jinja2>=3.1.0

# MCP support
mcp>=0.1.0

# LLM and Web Scraping
google-generativeai>=0.3.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
